refactor: 코드 리뷰 반영 (이름 오타 수정, constant 이동)
