+++ src/components/ErrorAlertFullScreen/NotFoundErrorAlertFullScreen/checkIsGetNotFound.ts
