+++ src/components/ErrorAlertFullScreen/NotFoundErrorAlertFullScreen/checkIsInvalidStatePage.ts
